requirements.txt
bs4
googlemaps
flake8
flake8-tabs
python-dateutil
pywin32
requests[security]
wxpython
#git+https://github.com/nstockton/speechlight.git#egg=speechlight
https://github.com/nstockton/speechlight/archive/master.zip
